% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AbsErr.R
\name{AbsErr}
\alias{AbsErr}
\title{Calculate the absolute error between forecast and observation}
\usage{
AbsErr(fcst, obs)
}
\arguments{
\item{fcst}{a N-vector representing N time instances of real-valued forecasts}

\item{obs}{a N-vector representing N time instances of real-valued observations}
}
\value{
numeric N-vector of absolute errors |fcst - obs|
}
\description{
Calculate the absolute error between forecast and observation
}
\examples{
fcst <- rnorm(20)
obs <- rnorm(20)
mean(AbsErr(fcst, obs))
}
\seealso{
SqErr, ScoreDiff, SkillScore
}

