% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Auc.R
\name{Auc}
\alias{Auc}
\title{Calculate area under the ROC curve (AUC) for a forecast and its verifying binary observation, and estimate the variance of the AUC}
\usage{
Auc(fcst, obs, handle.na = "na.fail")
}
\arguments{
\item{fcst}{vector of forecasts}

\item{obs}{vector of binary observations (0 for non-occurrence, 1 for occurrence of the event)}

\item{handle.na}{how should missing values in forecasts and observations be handled; possible values are 'na.fail' and 'only.complete.pairs'; default: 'na.fail'}
}
\value{
vector containing AUC and its estimated standard deviation
}
\description{
Calculate area under the ROC curve (AUC) for a forecast and its verifying binary observation, and estimate the variance of the AUC
}
\examples{
fcst <- runif(20)
obs <- rbinom(n=20, size=1, prob=fcst)
Auc(fcst, obs)
}
\references{
DeLong 1988
}
\seealso{
n/a
}

