% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{auc_cpp}
\alias{auc_cpp}
\title{Calculate AUC and its sampling standard deviation (Internal C++ implementation)}
\usage{
auc_cpp(fcst, obs)
}
\arguments{
\item{fcst}{numeric vector of forecasts (NAs are not allowed)}

\item{obs}{vector of binary observations (obs[t] evaluates to TRUE if event happens at instance t, to FALSE otherwise)}
}
\value{
AUC and its sampling standard deviation
}
\description{
Calculate AUC and its sampling standard deviation (Internal C++ implementation)
}
\seealso{
Auc AucDiff
}
