% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AucDiff.R
\name{AucDiff}
\alias{AucDiff}
\title{Calculate difference between areas under the ROC curve (AUC) between a forecast and a reference forecast for the same observation, and estimate the variance of the AUC difference}
\usage{
AucDiff(fcst, fcst.ref, obs, handle.na = "na.fail")
}
\arguments{
\item{fcst}{vector of forecasts}

\item{fcst.ref}{vector of reference forecasts}

\item{obs}{vector of binary observations (0 for non-occurrence, 1 for occurrence of the event)}

\item{handle.na}{how should missing values in forecasts and observations be handled; possible values are 'na.fail' and 'only.complete.triplets'; default: 'na.fail'}
}
\value{
vector with AUC difference, and estimated standard deviation
}
\description{
Calculate difference between areas under the ROC curve (AUC) between a forecast and a reference forecast for the same observation, and estimate the variance of the AUC difference
}
\examples{
fcst <- runif(20)
obs <- rbinom(n=20, size=1, prob=fcst)
fcst.ref <- rep(0.5, 20)
AucDiff(fcst, fcst.ref, obs)
}
\references{
DeLong et al (1988): Comparing the Areas under Two or More Correlated Receiver Operating Characteristic Curves: A Nonparametric Approach. Biometrics. http://dx.doi.org/10.2307/2531595
}
\seealso{
Auc
}

