% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aucdiff_cpp}
\alias{aucdiff_cpp}
\title{Calculate AUC difference `AUC(fcst,obs) - AUC(fcst_ref, obs)` of two forecasts for the same observations, and the sampling standard deviation of the AUC difference (Internal C++ implementation)}
\usage{
aucdiff_cpp(fcst, fcst_ref, obs)
}
\arguments{
\item{fcst}{numeric vector of forecasts (NAs are not allowed)}

\item{fcst_ref}{numeric vector of reference forecasts (NAs are not allowed)}

\item{obs}{vector of binary observations (obs[t] evaluates to TRUE if event happens at instance t, to FALSE otherwise)}
}
\value{
AUC values, their sampling standard deviations, the AUC difference, and their sampling standard deviations
}
\description{
Calculate AUC difference `AUC(fcst,obs) - AUC(fcst_ref, obs)` of two forecasts for the same observations, and the sampling standard deviation of the AUC difference (Internal C++ implementation)
}
\seealso{
Auc AucDiff
}
