% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimEns.R
\name{ClimEns}
\alias{ClimEns}
\title{Construct a climatological ensemble from a vector of observations.}
\usage{
ClimEns(obs, leave.one.out=FALSE)
}
\arguments{
\item{obs}{vector of length N. The observations.}

\item{leave.one.out}{logical, default=FALSE. If TRUE, the n-th observation is removed from the n-th row of the ensemble matrix.}
}
\value{
matrix with N rows and N-1 columns (if leave.one.out==TRUE) or N columns otherwise.
}
\description{
Construct a climatological ensemble from a vector of observations. Optionally, the climatological ensemble observation at time t can be constructed without the observation at time t (leave-one-out).
}
\examples{
data(eurotempforecast)
ClimEns(obs)
}
