% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dresscrps_cpp}
\alias{dresscrps_cpp}
\title{Dress CRPS}
\usage{
dresscrps_cpp(m, s, y)
}
\arguments{
\item{m}{vector of kernel means}

\item{s}{vector of kernel standard deviations}

\item{y}{observation}
}
\value{
crps
}
\description{
Dress CRPS
}
