% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreDiff.R
\name{DressCrpsDiff}
\alias{DressCrpsDiff}
\title{Calculate DressCrps Difference (deprecated, use function ScoreDiff instead)}
\usage{
DressCrpsDiff(dressed.ens, dressed.ens.ref, obs, probs = NA)
}
\arguments{
\item{dressed.ens}{the ensemble}

\item{dressed.ens.ref}{the reference ensemble}

\item{obs}{the observation}

\item{probs}{not used}
}
\value{
mean DressCrps difference
}
\description{
Calculate DressCrps Difference (deprecated, use function ScoreDiff instead)
}
\seealso{
ScoreDiff DressCrps DressEnsemble
}
