% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkillScore.R
\name{DressCrpss}
\alias{DressCrpss}
\title{Calculate DressCrps Skill Score (deprecated, use function SkillScore instead)}
\usage{
DressCrpss(dressed.ens, dressed.ens.ref, obs)
}
\arguments{
\item{dressed.ens}{the ensemble}

\item{dressed.ens.ref}{the reference ensemble}

\item{obs}{the observation}
}
\value{
DressCrps Skill Score
}
\description{
Calculate DressCrps Skill Score (deprecated, use function SkillScore instead)
}
\seealso{
SkillScore DressCrps DressEnsemble
}
