% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreDiff.R
\name{DressIgnDiff}
\alias{DressIgnDiff}
\title{Calculate DressIgn Difference (deprecated, use function ScoreDiff instead)}
\usage{
DressIgnDiff(dressed.ens, dressed.ens.ref, obs, probs = NA)
}
\arguments{
\item{dressed.ens}{the ensemble}

\item{dressed.ens.ref}{the reference ensemble}

\item{obs}{the observation}

\item{probs}{not used}
}
\value{
mean DressIgn difference
}
\description{
Calculate DressIgn Difference (deprecated, use function ScoreDiff instead)
}
\seealso{
ScoreDiff DressIgn
}
