% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreDiff.R
\name{EnsBrierDiff}
\alias{EnsBrierDiff}
\title{Calculate EnsBrier Difference (deprecated, use function ScoreDiff instead)}
\usage{
EnsBrierDiff(ens, ens.ref, obs, tau = NA, probs = NA)
}
\arguments{
\item{ens}{the ensemble}

\item{ens.ref}{the reference ensemble}

\item{obs}{the observation}

\item{tau}{not used}

\item{probs}{not used}
}
\value{
mean EnsBrier difference
}
\description{
Calculate EnsBrier Difference (deprecated, use function ScoreDiff instead)
}
\seealso{
ScoreDiff EnsBrier
}
