% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkillScore.R
\name{EnsBrierSs}
\alias{EnsBrierSs}
\title{Calculate EnsBrier Skill Score (deprecated, use function SkillScore instead)}
\usage{
EnsBrierSs(ens, ens.ref, obs, tau = NA)
}
\arguments{
\item{ens}{the ensemble}

\item{ens.ref}{the reference ensemble}

\item{obs}{the observation}

\item{tau}{not used}
}
\value{
EnsBrier skill score
}
\description{
Calculate EnsBrier Skill Score (deprecated, use function SkillScore instead)
}
\seealso{
SkillScore EnsBrier
}
