% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{enscrps_cpp}
\alias{enscrps_cpp}
\title{CRPS for ensemble forecasts (C++ implementation)}
\usage{
enscrps_cpp(ens, obs, R_new)
}
\arguments{
\item{ens}{Ensemble members as columns of a matrix}

\item{obs}{The verifying observations}

\item{R_new}{Size for ensemble adjustment}
}
\value{
vector of crps values
}
\description{
CRPS for ensemble forecasts (C++ implementation)
}
