% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreDiff.R
\name{EnsCrpsDiff}
\alias{EnsCrpsDiff}
\title{Calculate EnsCrps Difference (deprecated, use function ScoreDiff instead)}
\usage{
EnsCrpsDiff(ens, ens.ref, obs, probs = NA)
}
\arguments{
\item{ens}{the ensemble}

\item{ens.ref}{the reference ensemble}

\item{obs}{the observation}

\item{probs}{not used}
}
\value{
mean EnsCrps difference
}
\description{
Calculate EnsCrps Difference (deprecated, use function ScoreDiff instead)
}
\seealso{
ScoreDiff EnsCrps
}
