% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkillScore.R
\name{EnsCrpss}
\alias{EnsCrpss}
\title{Calculate EnsCrps Skill Score (deprecated, use function SkillScore instead)}
\usage{
EnsCrpss(ens, ens.ref, obs)
}
\arguments{
\item{ens}{the ensemble}

\item{ens.ref}{the reference ensemble}

\item{obs}{the observation}
}
\value{
EnsCrps skill score
}
\description{
Calculate EnsCrps Skill Score (deprecated, use function SkillScore instead)
}
\seealso{
SkillScore EnsCrps
}
