% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EnsQs.R
\name{EnsQs}
\alias{EnsQs}
\title{Calculate the ensemble-adjusted Quadratic Score (QS) for categorical forecasts}
\usage{
EnsQs(ens, obs, R.new = NA)
}
\arguments{
\item{ens}{a N*K matrix representing N time instances of categorical ensemble forecasts; ens[t,k] indicates the number of ensemble members that predict category k at time t}

\item{obs}{a N*K matrix, with obs[i,k] either 0 or 1, and all rowSums equal to one; y[t,k] = 1 if category k verifies at time t, otherwise y[t,k] = 0}

\item{R.new}{ensemble size for which the scores should be adjusted}
}
\value{
numeric vector of length N with the ensemble-adjusted quadratic score values
}
\description{
Calculate the ensemble-adjusted Quadratic Score (QS) for categorical forecasts
}
\examples{
# N=5 time instances, R=10 ensemble members, K=3 classes
ens <- t(rmultinom(n=5, size=10, prob=c(.3, .1, .6)))
obs <- t(rmultinom(n=5, size=1,  prob=c(.3, .1, .6)))
EnsQs(ens, obs, R.new=Inf)
}
\references{
n/a
}
\seealso{
EnsBrier, EnsRps, EnsCrps
}

