% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsRps.R
\name{EnsRps}
\alias{EnsRps}
\title{Calculate the ensemble-adjusted Ranked Probability Score (RPS) for categorical forecasts}
\usage{
EnsRps(ens, obs, R.new = NA)
}
\arguments{
\item{ens}{a N*R matrix representing N time instances of categorical ensemble forecasts; ens[t,r] indicates the category that the r-th ensemble member predicts for time t}

\item{obs}{a vector of length N, with obs[t] is the category observed at time t}

\item{R.new}{ensemble size for which the scores should be adjusted}
}
\value{
numeric vector of length N with the ensemble-adjusted RPS values
}
\description{
Calculate the ensemble-adjusted Ranked Probability Score (RPS) for categorical forecasts
}
\examples{
data(eurotempforecast)
EnsRps(ens.cat, obs.cat, R.new=Inf)
}
\references{
n/a
}
\seealso{
EnsBrier, EnsQs, EnsCrps
}

