% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkillScore.R
\name{EnsRpss}
\alias{EnsRpss}
\title{Calculate EnsRps Skill Score (deprecated, use function SkillScore instead)}
\usage{
EnsRpss(ens, ens.ref, obs, format = c("category", "members"))
}
\arguments{
\item{ens}{the ensemble}

\item{ens.ref}{the reference ensemble}

\item{obs}{the observation}

\item{format}{see `EnsRps`}
}
\value{
EnsRps skill score
}
\description{
Calculate EnsRps Skill Score (deprecated, use function SkillScore instead)
}
\seealso{
SkillScore EnsRps
}
