% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eurotempforecast.R
\docType{data}
\name{eurotempforecast}
\alias{eurotempforecast}
\title{Seasonal ensemble forecast of European average summer temperature}
\format{Variables `ens` and `obs`
\itemize{
  \item `ens` ensemble data
  \item `obs` observation data
}}
\usage{
data(eurotempforecast)
}
\description{
A dataset containing ...
}
\keyword{datasets}

