% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkillScore.R
\name{FairBrierSs}
\alias{FairBrierSs}
\title{Calculate FairBrier Skill Score (deprecated, use function SkillScore instead)}
\usage{
FairBrierSs(ens, ens.ref, obs, tau = NA)
}
\arguments{
\item{ens}{the ensemble}

\item{ens.ref}{the reference ensemble}

\item{obs}{the observation}

\item{tau}{not used}
}
\value{
FairBrier skill score
}
\description{
Calculate FairBrier Skill Score (deprecated, use function SkillScore instead)
}
\seealso{
SkillScore EnsBrier
}
