% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreDiff.R
\name{FairCrpsDiff}
\alias{FairCrpsDiff}
\title{Calculate FairCrps Difference (deprecated, use function ScoreDiff instead)}
\usage{
FairCrpsDiff(ens, ens.ref, obs, probs = NA)
}
\arguments{
\item{ens}{the ensemble}

\item{ens.ref}{the reference ensemble}

\item{obs}{the observation}

\item{probs}{not used}
}
\value{
mean FairCrps difference
}
\description{
Calculate FairCrps Difference (deprecated, use function ScoreDiff instead)
}
\seealso{
ScoreDiff EnsCrps
}
