% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkillScore.R
\name{FairCrpss}
\alias{FairCrpss}
\title{Calculate FairCrps Skill Score (deprecated, use function SkillScore instead)}
\usage{
FairCrpss(ens, ens.ref, obs)
}
\arguments{
\item{ens}{the ensemble}

\item{ens.ref}{the reference ensemble}

\item{obs}{the observation}
}
\value{
FairCrps skill score
}
\description{
Calculate FairCrps Skill Score (deprecated, use function SkillScore instead)
}
\seealso{
SkillScore EnsCrps
}
