% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SkillScore.R
\name{FairRpss}
\alias{FairRpss}
\title{Calculate FairRps Skill Score (deprecated, use function SkillScore instead)}
\usage{
FairRpss(ens, ens.ref, obs)
}
\arguments{
\item{ens}{the ensemble}

\item{ens.ref}{the reference ensemble}

\item{obs}{the observation}
}
\value{
FairRps skill score
}
\description{
Calculate FairRps Skill Score (deprecated, use function SkillScore instead)
}
\seealso{
SkillScore EnsRps
}

