% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreDiff.R
\name{GaussCrpsDiff}
\alias{GaussCrpsDiff}
\title{Calculate GaussCrps Difference (deprecated, use function ScoreDiff instead)}
\usage{
GaussCrpsDiff(mean, sd, mean.ref, sd.ref, obs, probs = NA)
}
\arguments{
\item{mean}{forecast means}

\item{sd}{forecast standard deviations}

\item{mean.ref}{reference forecast means}

\item{sd.ref}{reference forecast standard deviations}

\item{obs}{the observation}

\item{probs}{not used}
}
\value{
mean GaussCrps difference
}
\description{
Calculate GaussCrps Difference (deprecated, use function ScoreDiff instead)
}
\seealso{
ScoreDiff GaussCrps
}
