% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkillScore.R
\name{GaussCrpss}
\alias{GaussCrpss}
\title{Calculate GaussCrps Skill Score (deprecated, use function SkillScore instead)}
\usage{
GaussCrpss(mean, sd, mean.ref, sd.ref, obs)
}
\arguments{
\item{mean}{forecast means}

\item{sd}{forecast standard deviations}

\item{mean.ref}{reference forecast means}

\item{sd.ref}{reference forecast standard deviations}

\item{obs}{the observation}
}
\value{
GaussCrps skill score
}
\description{
Calculate GaussCrps Skill Score (deprecated, use function SkillScore instead)
}
\seealso{
SkillScore GaussCrps
}
