% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SkillScore.R
\name{SkillScore}
\alias{SkillScore}
\title{Calculate a skill score and assess uncertainty}
\usage{
SkillScore(scores, scores.ref, N.eff = NA, score.perf = 0,
  handle.na = "na.fail")
}
\arguments{
\item{scores}{vector of verification scores}

\item{scores.ref}{vector of verification scores of the reference forecast, must be of the same length as `scores`}

\item{N.eff}{user-defined effective sample size to be used to estimate the sampling uncertainty; if NA, the length of `scores` is used; default: NA}

\item{score.perf}{a numeric constant, indicating the value that the score would assign to the perfect forecast}

\item{handle.na}{how should missing values in scores vectors be handled; possible values are 'na.fail' and 'use.pairwise.complete'; default: 'na.fail'}
}
\value{
vector with skill score and its estimated standard deviation
}
\description{
Calculate a skill score and assess uncertainty
}
\examples{
fcst <- rnorm(20)
fcst.ref <- rnorm(20)
obs <- rnorm(20)
SkillScore(SqErr(fcst, obs), SqErr(fcst.ref, obs))
}
\references{
n/a
}
\seealso{
n/a
}

