% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecsVerification.R
\docType{package}
\name{SpecsVerification}
\alias{SpecsVerification}
\title{SpecsVerification - Forecast verification routines}
\description{
SpecsVerification - Forecast verification routines
}
\author{
: Stefan Siegert
}
