% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsCrps.R
\name{EnsCrps}
\alias{EnsCrps}
\title{Calculate the ensemble-adjusted Continuous Ranked Probability Score (CRPS)}
\usage{
EnsCrps(ens, obs, R.new = NA)
}
\arguments{
\item{ens}{a N*R representing N time instances of real-valued R-member ensemble forecasts}

\item{obs}{a numeric vector of length N with real-valued observations}

\item{R.new}{ensemble size for which the scores should be adjusted}
}
\value{
numeric vector of length N with the ensemble-adjusted CRPS values
}
\description{
Calculate the ensemble-adjusted Continuous Ranked Probability Score (CRPS)
}
\examples{
ens <- matrix(rnorm(100), 20, 5)
obs <- rnorm(20)
mean(EnsCrps(ens, obs, R.new=Inf))
}
\references{
Ferro CAT, Richardson SR, Weigel AP (2008) On the effect of ensemble size on the discrete and continuous ranked probability scores. Meteorological Applications. doi: 10.1002/met.45
}
\seealso{
EnsBrier, EnsRps
}

