% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreDiff.R
\name{EnsRpsDiff}
\alias{EnsRpsDiff}
\title{Calculate EnsRps Difference (deprecated, use function ScoreDiff instead)}
\usage{
EnsRpsDiff(ens, ens.ref, obs, probs = NA, format = c("category", "members"))
}
\arguments{
\item{ens}{the ensemble}

\item{ens.ref}{the reference ensemble}

\item{obs}{the observation}

\item{probs}{not used}

\item{format}{see `EnsRps`}
}
\value{
mean EnsRps difference
}
\description{
Calculate EnsRps Difference (deprecated, use function ScoreDiff instead)
}
\seealso{
ScoreDiff EnsRps
}
