% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqErr.R
\name{SqErr}
\alias{SqErr}
\title{Calculate the squared error between forecast and observation}
\usage{
SqErr(fcst, obs)
}
\arguments{
\item{fcst}{a N-vector representing N time instances of real-valued forecasts}

\item{obs}{a N-vector representing N time instances of real-valued observations}
}
\value{
numeric N-vector of squared errors
}
\description{
Calculate the squared error between forecast and observation
}
\examples{
data(eurotempforecast)
mean(SqErr(rowMeans(ens), obs))
}
\seealso{
AbsErr, ScoreDiff, SkillScore
}
